Moving the CSV files to the int-testing area. Saving the AT changes and removal of Async SQS Client Bean
