+++ app/src/main/java/com/ebsolutions/papertrail/financialdataingestionservice/config/EventConfig.java
@@ -8,28 +8,18 @@ import org.springframework.context.annotation.Profile;
 import software.amazon.awssdk.auth.credentials.AwsBasicCredentials;
 import software.amazon.awssdk.auth.credentials.StaticCredentialsProvider;
 import software.amazon.awssdk.regions.Region;
-import software.amazon.awssdk.services.sqs.SqsAsyncClient;
 import software.amazon.awssdk.services.sqs.SqsClient;
 
 @Configuration
 public class EventConfig {
-  @Value("${infrastructure.endpoint:`Database endpoint not found in environment`}")
+  @Value("${infrastructure.endpoint:`Infrastructure endpoint not found in environment`}")
   protected String endpoint;
-
 
   @Bean
   @Profile({"local", "default"})
   public SqsClient localSqsClientInstantiation() {
     return SqsClient.builder()
         .region(Region.US_EAST_1)
-        .endpointOverride(URI.create(endpoint))
-        .credentialsProvider(staticCredentialsProvider())
-        .build();
-  }
-
-  @Bean
-  public SqsAsyncClient sqsAsyncClient() {
-    return SqsAsyncClient.builder().region(Region.US_EAST_1)
         .endpointOverride(URI.create(endpoint))
         .credentialsProvider(staticCredentialsProvider())
         .build();
